ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
google-cloud-speech>=2.20.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
